downloadSPFileFromApp
This code is basically used in SharePoint Online for downloading the file from document library using SharePoint App. This is one of the ways to achieve this functionality without using O365 licensed user.
Using this App ID concept for doing similar kind of operations in SharePoint Online will reduce costing since it just needs an App registration and not any O365 license.
I have created this simple console application which can be used and modified as per need.